consolidate photoobj
