--- pydl/photoop/photoobj/__init__.py
@@ -1,7 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the photoobj directory of photoop.
-"""
-from .sdss_calibv import sdss_calibv
-from .unwrap_objid import unwrap_objid
