+++ pydl/photoop/tests/__init__.py
@@ -1,5 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
 """
-This is the pydl/photoop/photoobj/tests directory.
+This is the pydl/photoop/tests directory.
 """
